Avoid warnings
