+++ .eslintrc.json
@@ -37,7 +37,7 @@
   "overrides": [
     {
       "files": [
-        "**/test/**/*.ts",
+        "**/test/**/*.{ts,js}",
         "*.spec.ts"
       ],
       "env": {
