+++ packages/cache/inmemory-lru/package.json
@@ -8,12 +8,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  "devDependencies": {
-    
-    "ts-jest": "25.3.1",
-    "ts-node-dev": "1.0.0-pre.44",
-    "typescript": "3.8.3"
-  },
   "sideEffects": false,
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
@@ -24,8 +18,7 @@
   "dependencies": {
     "@graphql-mesh/types": "0.0.20",
     "@types/lru-cache": "5.1.0",
-    "lru-cache": "5.1.1",
-    "ts-node": "8.8.2"
+    "lru-cache": "5.1.1"
   },
   "publishConfig": {
     "access": "public",
