+++ packages/cli/package.json
@@ -40,7 +40,6 @@
     "@graphql-toolkit/graphql-file-loader": "0.10.3",
     "apollo-server": "2.12.0",
     "chalk": "4.0.0",
-    "change-case": "4.1.1",
     "js-yaml": "3.13.1",
     "mkdirp": "1.0.4",
     "pascal-case": "3.1.1",
@@ -50,7 +49,6 @@
   "devDependencies": {
     "@types/js-yaml": "3.12.3",
     "@types/mkdirp": "1.0.0",
-    
     "@types/yargs": "15.0.4"
   },
   "publishConfig": {
