+++ packages/handlers/graphql/package.json
@@ -15,7 +15,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  
   "dependencies": {
     "@graphql-mesh/types": "0.0.20",
     "@graphql-toolkit/core": "0.10.3",
