+++ packages/handlers/grpc/package.json
@@ -17,14 +17,13 @@
     "grpc": "*"
   },
   "devDependencies": {
-    
     "grpc": "1.24.2"
   },
   "dependencies": {
     "@graphql-mesh/types": "0.0.20",
     "graphql-scalars": "1.1.0",
     "grpc-caller": "0.13.0",
-    "iter-tools": "^6.2.6",
+    "iter-tools": "6.2.6",
     "protobufjs": "6.8.9"
   },
   "publishConfig": {
