+++ packages/handlers/json-schema/package.json
@@ -16,7 +16,6 @@
     "graphql": "*"
   },
   "devDependencies": {
-    
     "@types/url-join": "4.0.0"
   },
   "dependencies": {
@@ -24,7 +23,6 @@
     "@graphql-mesh/types": "0.0.20",
     "@graphql-mesh/utils": "0.0.20",
     "aggregate-error": "3.0.1",
-    "change-case": "4.1.1",
     "fetchache": "0.0.3",
     "graphql-scalars": "1.1.0",
     "pascal-case": "3.1.1",
