+++ packages/handlers/mongoose/src/index.ts
@@ -1,27 +1,12 @@
-import {
-  composeWithMongoose,
-  composeWithMongooseDiscriminators
-} from 'graphql-compose-mongoose';
-import {
-  SchemaComposer,
-  ObjMap,
-  ObjectTypeComposerFieldConfigDefinition
-} from 'graphql-compose';
+import { composeWithMongoose, composeWithMongooseDiscriminators } from 'graphql-compose-mongoose';
+import { SchemaComposer, ObjMap, ObjectTypeComposerFieldConfigDefinition } from 'graphql-compose';
 import { MeshHandlerLibrary, YamlConfig } from '@graphql-mesh/types';
-import { camelCase } from 'change-case';
+import { camelCase } from 'camel-case';
 import { ArgsMap } from 'graphql-compose/lib/ObjectTypeComposer';
 import mongoose from 'mongoose';
 import { isAbsolute, join } from 'path';
 
-const modelQueryOperations = [
-  'findById',
-  'findByIds',
-  'findOne',
-  'findMany',
-  'count',
-  'connection',
-  'pagination'
-];
+const modelQueryOperations = ['findById', 'findByIds', 'findOne', 'findMany', 'count', 'connection', 'pagination'];
 
 const modelMutationOperations = [
   'createOne',
@@ -31,17 +16,12 @@ const modelMutationOperations = [
   'updateMany',
   'removeById',
   'removeOne',
-  'removeMany'
+  'removeMany',
 ];
 
-async function loadExternalModuleByExpression(
-  expression: string,
-  defaultExportName?: string
-) {
+async function loadExternalModuleByExpression(expression: string, defaultExportName?: string) {
   const [modulePath, exportName = defaultExportName] = expression.split('#');
-  const absoluteModulePath = isAbsolute(modulePath)
-    ? modulePath
-    : join(process.cwd(), modulePath);
+  const absoluteModulePath = isAbsolute(modulePath) ? modulePath : join(process.cwd(), modulePath);
   const mod = await import(absoluteModulePath);
 
   return (exportName && mod[exportName]) || mod.default || mod;
@@ -49,47 +29,30 @@ async function loadExternalModuleByExpression(
 
 const handler: MeshHandlerLibrary<YamlConfig.MongooseHandler> = {
   async getMeshSource({ config, hooks }) {
-
     const schemaComposer = new SchemaComposer();
 
     if (config.connectionString) {
       await mongoose.connect(config.connectionString, {
         useNewUrlParser: true,
-        useUnifiedTopology: true
+        useUnifiedTopology: true,
       });
     }
 
-    const queryFields: ObjMap<ObjectTypeComposerFieldConfigDefinition<
-      any,
-      any,
-      ArgsMap
-    >> = {};
-    const mutationFields: ObjMap<ObjectTypeComposerFieldConfigDefinition<
-      any,
-      any,
-      ArgsMap
-    >> = {};
+    const queryFields: ObjMap<ObjectTypeComposerFieldConfigDefinition<any, any, ArgsMap>> = {};
+    const mutationFields: ObjMap<ObjectTypeComposerFieldConfigDefinition<any, any, ArgsMap>> = {};
 
     if (config.models) {
       await Promise.all(
         config.models.map(async modelConfig => {
-          const model = await loadExternalModuleByExpression(
-            modelConfig.path,
-            modelConfig.name
-          );
-          const modelTC = composeWithMongoose(
-            model,
-            modelConfig.options as any
-          );
+          const model = await loadExternalModuleByExpression(modelConfig.path, modelConfig.name);
+          const modelTC = composeWithMongoose(model, modelConfig.options as any);
           for (const queryOperation of modelQueryOperations) {
-            queryFields[
-              camelCase(`${modelConfig.name}_${queryOperation}`)
-            ] = modelTC.getResolver(queryOperation);
+            queryFields[camelCase(`${modelConfig.name}_${queryOperation}`)] = modelTC.getResolver(queryOperation);
           }
           for (const mutationOperation of modelMutationOperations) {
-            mutationFields[
-              camelCase(`${modelConfig.name}_${mutationOperation}`)
-            ] = modelTC.getResolver(mutationOperation);
+            mutationFields[camelCase(`${modelConfig.name}_${mutationOperation}`)] = modelTC.getResolver(
+              mutationOperation
+            );
           }
         })
       );
@@ -102,19 +65,16 @@ const handler: MeshHandlerLibrary<YamlConfig.MongooseHandler> = {
             discriminatorConfig.path,
             discriminatorConfig.name
           );
-          const discriminatorTC = composeWithMongooseDiscriminators(
-            discriminator,
-            discriminatorConfig.options as any
-          );
+          const discriminatorTC = composeWithMongooseDiscriminators(discriminator, discriminatorConfig.options as any);
           for (const queryOperation of modelQueryOperations) {
-            queryFields[
-              camelCase(`${discriminatorConfig.name}_${queryOperation}`)
-            ] = discriminatorTC.getResolver(queryOperation);
+            queryFields[camelCase(`${discriminatorConfig.name}_${queryOperation}`)] = discriminatorTC.getResolver(
+              queryOperation
+            );
           }
           for (const mutationOperation of modelMutationOperations) {
-            mutationFields[
-              camelCase(`${discriminatorConfig.name}_${mutationOperation}`)
-            ] = discriminatorTC.getResolver(mutationOperation);
+            mutationFields[camelCase(`${discriminatorConfig.name}_${mutationOperation}`)] = discriminatorTC.getResolver(
+              mutationOperation
+            );
           }
         })
       );
@@ -127,9 +87,9 @@ const handler: MeshHandlerLibrary<YamlConfig.MongooseHandler> = {
     hooks.on('destroy', () => mongoose.disconnect());
 
     return {
-      schema
+      schema,
     };
-  }
+  },
 };
 
 export default handler;
