+++ packages/handlers/odata/package.json
@@ -18,7 +18,6 @@
   "devDependencies": {
     "@types/graphql-fields": "1.3.2",
     "@types/jsdom": "16.2.0",
-    
     "@types/url-join": "4.0.0"
   },
   "dependencies": {
