+++ packages/runtime/package.json
@@ -8,7 +8,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  
   "sideEffects": false,
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
