+++ packages/transforms/cache/package.json
@@ -16,7 +16,6 @@
     "graphql": "*"
   },
   "devDependencies": {
-    
     "@types/object-hash": "1.3.1"
   },
   "dependencies": {
