+++ packages/transforms/cache/src/index.ts
@@ -6,26 +6,9 @@ import {
   ResolversComposerMapping,
   ResolversComposition,
 } from '@graphql-toolkit/common';
-import { Interpolator } from '@ardatan/string-interpolation';
-import { format } from 'date-fns';
-import objectHash from 'object-hash';
+import { computeCacheKey } from './compute-cache-key';
 
-const interpolator = new Interpolator({
-  delimiter: ['{', '}'],
-});
-interpolator.addAlias('typeName', 'info.parentType.name');
-interpolator.addAlias('type', 'info.parentType.name');
-interpolator.addAlias('parentType', 'info.parentType.name');
-interpolator.addAlias('fieldName', 'info.fieldName');
-interpolator.registerModifier('date', (formatStr: string) => format(new Date(), formatStr));
-interpolator.registerModifier('hash', (value: any) => objectHash(value, { ignoreUnknown: true }));
-
-export const cacheTransform: TransformFn<YamlConfig.CacheTransformConfig[]> = async ({
-  config,
-  schema,
-  cache,
-  hooks,
-}) => {
+const cacheTransform: TransformFn<YamlConfig.CacheTransformConfig[]> = async ({ config, schema, cache, hooks }) => {
   // We need to use `schemaReady` hook and not the schema directly because we need to make sure to run cache after all
   // other transformations are done, and to make sure that custom resolve are already loaded and merged into the schema
   hooks.on('schemaReady', ({ schema, applyResolvers }) => {
@@ -86,25 +69,5 @@ export const cacheTransform: TransformFn<YamlConfig.CacheTransformConfig[]> = as
   });
   return schema;
 };
-
-export function computeCacheKey(options: {
-  keyStr: string | undefined;
-  args: Record<string, any>;
-  info: GraphQLResolveInfo;
-}): string {
-  const argsHash = options.args ? objectHash(options.args, { ignoreUnknown: true }) : '';
-
-  if (!options.keyStr) {
-    return `${options.info.parentType.name}-${options.info.fieldName}-${argsHash}`;
-  }
-
-  const templateData = {
-    args: options.args,
-    argsHash,
-    info: options.info || null,
-  };
-
-  return interpolator.parse(options.keyStr, templateData);
-}
 
 export default cacheTransform;
