+++ packages/transforms/cache/test/cache.spec.ts
@@ -2,7 +2,8 @@ import { YamlConfig, Hooks } from '@graphql-mesh/types';
 import { InMemoryLRUCache } from '@graphql-mesh/cache-inmemory-lru';
 import { addResolversToSchema } from 'graphql-tools-fork';
 import { GraphQLSchema, buildSchema, execute, parse, DocumentNode } from 'graphql';
-import cacheTransform, { computeCacheKey } from '../src';
+import cacheTransform from '../src';
+import { computeCacheKey } from '../src/compute-cache-key';
 import objectHash from 'object-hash';
 import { format } from 'date-fns';
 import { applyResolversHooksToSchema } from '@graphql-mesh/runtime';
