+++ packages/transforms/extend/package.json
@@ -15,13 +15,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  "devDependencies": {
-    
-    "jest": "25.2.7",
-    "ts-jest": "25.3.1",
-    "ts-node-dev": "1.0.0-pre.44",
-    "typescript": "3.8.3"
-  },
   "dependencies": {
     "@graphql-mesh/types": "0.0.20"
   },
