+++ packages/transforms/rename/package.json
@@ -15,7 +15,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  
   "dependencies": {
     "@graphql-mesh/types": "0.0.20",
     "graphql-tools-fork": "8.9.4"
