+++ packages/transforms/rename/test/rename.spec.ts
@@ -1,4 +1,4 @@
-import { renameTransform } from './../src/index';
+import renameTransform from './../src/index';
 import { buildSchema, printSchema } from 'graphql';
 import { InMemoryLRUCache } from '@graphql-mesh/cache-inmemory-lru';
 import { Hooks } from '@graphql-mesh/types';
@@ -20,8 +20,8 @@ describe('rename', () => {
       config: [
         {
           from: 'User',
-          to: 'MyUser'
-        }
+          to: 'MyUser',
+        },
       ],
       cache: new InMemoryLRUCache(),
       hooks: new Hooks(),
