+++ packages/types/package.json
@@ -15,7 +15,6 @@
   "peerDependencies": {
     "graphql": "*"
   },
-  
   "dependencies": {
     "fetchache": "0.0.3",
     "tsee": "1.3.0"
